Move some debug functions (which I'll likely completely delete later).
