+++ meson.build
@@ -176,6 +176,7 @@ EMU_SRC = [
 	'src/config_manage.cpp',
 	'src/PATCHES/rom.cpp',
 	'src/UTIL/date_to_seconds.cpp',
+	'src/UTIL/dbglog.cpp',
 	'src/LANG/intl_chars.cpp',
 	'lib/tomlc99/toml.c',
 ]
