+++ src/UTIL/dbglog.cpp
@@ -1,0 +1,114 @@
+/*
+	dbglog.cpp
+
+    Based on GLOBGLUE.c.
+
+	Copyright (C) 2003 Bernd Schmidt, Philip Cummins, Paul C. Pratt
+
+	You can redistribute this file and/or modify it under the terms
+	of version 2 of the GNU General Public License as published by
+	the Free Software Foundation.  You should have received a copy
+	of the license along with this file; see the file COPYING.
+
+	This file is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+	license for more details.
+*/
+
+#include "sys_dependencies.h"
+
+#include "UI/my_os_glue.h"
+#include "UTIL/endian.h"
+#include "EMCONFIG.h"
+#include "dbglog.h"
+
+#if WantDisasm
+extern void m68k_WantDisasmContext(void);
+
+void dbglog_StartLine(void)
+{
+	m68k_WantDisasmContext();
+	dbglog_writeCStr(" ");
+}
+#endif
+
+#if dbglog_HAVE
+void dbglog_WriteMemArrow(bool WriteMem)
+{
+	if (WriteMem) {
+		dbglog_writeCStr(" <- ");
+	} else {
+		dbglog_writeCStr(" -> ");
+	}
+}
+
+void dbglog_AddrAccess(char *s, uint32_t Data,
+	bool WriteMem, uint32_t addr)
+{
+	dbglog_StartLine();
+	dbglog_writeCStr(s);
+	dbglog_writeCStr("[");
+	dbglog_writeHex(addr);
+	dbglog_writeCStr("]");
+	dbglog_WriteMemArrow(WriteMem);
+	dbglog_writeHex(Data);
+	dbglog_writeReturn();
+}
+void dbglog_Access(char *s, uint32_t Data, bool WriteMem)
+{
+	dbglog_StartLine();
+	dbglog_writeCStr(s);
+	dbglog_WriteMemArrow(WriteMem);
+	dbglog_writeHex(Data);
+	dbglog_writeReturn();
+}
+
+void dbglog_WriteNote(char *s)
+{
+	dbglog_StartLine();
+	dbglog_writeCStr(s);
+	dbglog_writeReturn();
+}
+
+void dbglog_WriteSetBool(char *s, bool v)
+{
+	dbglog_StartLine();
+	dbglog_writeCStr(s);
+	dbglog_writeCStr(" <- ");
+	if (v) {
+		dbglog_writeCStr("1");
+	} else {
+		dbglog_writeCStr("0");
+	}
+	dbglog_writeReturn();
+}
+#endif
+
+#if WantAbnormalReports
+static bool GotOneAbnormal = false;
+#endif
+
+#if WantAbnormalReports
+void DoReportAbnormalID(uint16_t id
+#if dbglog_HAVE
+	, char *s
+#endif
+	)
+{
+#if dbglog_HAVE
+	dbglog_StartLine();
+	dbglog_writeCStr("*** abnormal : ");
+	dbglog_writeCStr(s);
+	dbglog_writeReturn();
+#endif
+
+	if (! GotOneAbnormal) {
+		WarnMsgAbnormalID(id);
+#if ReportAbnormalInterrupt
+		SetInterruptButton(true);
+#endif
+		GotOneAbnormal = true;
+	}
+}
+#endif
